+++ src/inventorum.ebay/inventorum/ebay/apps/products/tests/test_batch_publish.py
@@ -7,7 +7,6 @@ import mock
 from inventorum.ebay.apps.products.tests import ProductTestMixin
 from inventorum.ebay.tests import StagingTestAccount, ApiTest
 from inventorum.ebay.tests.testcases import EbayAuthenticatedAPITestCase
-from inventorum.util.tests import celery_test_case
 
 log = logging.getLogger(__name__)
 
@@ -30,11 +29,12 @@ class TestBatchPublish(EbayAuthenticatedAPITestCase, ProductTestMixin):
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response.data['error']['key'], 'multiple.errors')
         self.assertEqual(response.data['error']['detail'][self.product_1.inv_id], ['You need to select category'])
-        self.assertEqual(response.data['error']['detail'][StagingTestAccount.Products.PRODUCT_VALID_FOR_PUBLISHING], ['Neither product or account have configured shipping '
-                                                               'services'])
+        self.assertEqual(response.data['error']['detail'][StagingTestAccount.Products.PRODUCT_VALID_FOR_PUBLISHING],
+                         ['Neither product or account have configured shipping services'])
 
     @mock.patch('inventorum.ebay.apps.products.resources.schedule_ebay_item_publish')
-    def test_publish(self, mock):
+    @mock.patch('inventorum.ebay.apps.products.services.PublishingService.initialize_publish_attempt')
+    def test_publish(self, schedule_ebay_item_publish_mock, initialize_publish_attempt_mock):
         product_2 = self.get_product(StagingTestAccount.Products.PRODUCT_VALID_FOR_PUBLISHING, self.account)
         self.assign_valid_shipping_services(self.product_1)
         self.assign_valid_shipping_services(product_2)
@@ -51,4 +51,5 @@ class TestBatchPublish(EbayAuthenticatedAPITestCase, ProductTestMixin):
         log.debug('Got response: %s', response)
         self.assertEqual(response.status_code, 200)
 
-        self.assertEqual(mock.call_count, 2)
+        self.assertEqual(initialize_publish_attempt_mock.call_count, 2)
+        self.assertEqual(schedule_ebay_item_publish_mock.call_count, 2)