+++ CHANGES.rst
@@ -4,6 +4,8 @@ CHANGES
 
 develop
 =======
+- INV-4631/INV-4457 Avoid double publishing by using database locks in the resource
+
 2015-06-16 0.1.6
 ================
 - add ``rc`` environment for release candidate environment